Fix high-severity bugs in syscalls, block device, memfs, HPET, MSI
- sys_poll: cap fd count at 1024 to prevent unbounded heap allocation
- Block device: group cache misses into contiguous sub-ranges for reads
- memfs remove_dir: fix inverted kind check (was rejecting dirs, accepting files)
- HPET ticks_to_nanos: use u128 intermediate to prevent overflow
- MSI: mask LAPIC ID to 8 bits for xAPIC address format
- sys_shm_map: validate addr_hint alignment and user-space range
- MAP_PHYSICAL: validate phys_addr alignment, rollback on partial map failure

+++ kernel/src/drivers/hpet/instant.rs
@@ -13,7 +13,8 @@ pub struct HpetTimer {
 impl HpetTimer {
     pub fn ticks_to_nanos(&self, ticks: u64) -> u64 {
         // Convert femtoseconds to nanoseconds: divide by 1,000,000
-        (ticks * self.frequency) / 1_000_000
+        // Use u128 to avoid overflow for large tick deltas
+        ((ticks as u128 * self.frequency as u128) / 1_000_000) as u64
     }
 
     pub fn nanos_to_ticks(&self, nanos: u128) -> u64 {
